requirements.txt
openai-whisper
torch
tqdm
soundfile
numpy
transformers
pydantic
